+++ src/TestSuiteStorage/Fixture/Dataset/EntityC.php
@@ -1,0 +1,10 @@
+<?php
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\TestSuite\Storage\Fixture\Dataset;
+
+use Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract;
+
+class EntityC extends DatasetEntityContract
+{
+}